LayoutTests/imported/w3c/web-platform-tests/pointerevents/deviceproperties/pointer-event-has-device-properties-uniqueid-from-pointer-event-init.tentative-expected.txt


FAIL PointerEvent.deviceProperties via DevicePropertiesInit promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: DeviceProperties"
FAIL PointerEvent.deviceProperties via empty DevicePropertiesInit promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: DeviceProperties"
FAIL No deviceProperties in PointerEventInit assert_equals: expected (object) null but got (undefined) undefined
